+++ tap_echo.c
@@ -78,14 +78,14 @@ typedef struct {
 	LADSPA_Data * input_R;
 	LADSPA_Data * output_R;
 
-	unsigned long sample_rate;
+	int sample_rate;
 	LADSPA_Data mpx_out_L;
 	LADSPA_Data mpx_out_R;
 
 	LADSPA_Data * ringbuffer_L;
 	LADSPA_Data * ringbuffer_R;
-	unsigned long * buffer_pos_L;
-	unsigned long * buffer_pos_R;
+	int * buffer_pos_L;
+	int * buffer_pos_R;
 
 	LADSPA_Data run_adding_gain;
 } Echo;
@@ -113,9 +113,9 @@ instantiate_Echo(const LADSPA_Descriptor * Descriptor,
 		     calloc(MAX_DELAY * ((Echo *)ptr)->sample_rate / 1000,
 			    sizeof(LADSPA_Data))) == NULL)
 			exit(1);
-		if ((((Echo *)ptr)->buffer_pos_L = calloc(1, sizeof(unsigned long))) == NULL)
+		if ((((Echo *)ptr)->buffer_pos_L = calloc(1, sizeof(int))) == NULL)
 			exit(1);
-		if ((((Echo *)ptr)->buffer_pos_R = calloc(1, sizeof(unsigned long))) == NULL)
+		if ((((Echo *)ptr)->buffer_pos_R = calloc(1, sizeof(int))) == NULL)
 			exit(1);
 		
 		*(((Echo *)ptr)->buffer_pos_L) = 0;
@@ -220,7 +220,7 @@ run_Echo(LADSPA_Handle Instance,
 	 unsigned long SampleCount) {
 	
 	Echo * ptr;
-	unsigned long sample_index;
+	int sample_index;
 
 	LADSPA_Data delaytime_L;
 	LADSPA_Data delaytime_R;
@@ -238,9 +238,9 @@ run_Echo(LADSPA_Handle Instance,
 	LADSPA_Data * input_R;
 	LADSPA_Data * output_R;
 
-	unsigned long sample_rate;
-	unsigned long buflen_L;
-	unsigned long buflen_R;
+	int sample_rate;
+	int buflen_L;
+	int buflen_R;
 
 	LADSPA_Data out_L = 0;
 	LADSPA_Data out_R = 0;
@@ -327,7 +327,7 @@ run_adding_gain_Echo(LADSPA_Handle Instance,
 	 unsigned long SampleCount) {
 	
 	Echo * ptr;
-	unsigned long sample_index;
+	int sample_index;
 
 	LADSPA_Data delaytime_L;
 	LADSPA_Data delaytime_R;
@@ -345,9 +345,9 @@ run_adding_gain_Echo(LADSPA_Handle Instance,
 	LADSPA_Data * input_R;
 	LADSPA_Data * output_R;
 
-	unsigned long sample_rate;
-	unsigned long buflen_L;
-	unsigned long buflen_R;
+	int sample_rate;
+	int buflen_L;
+	int buflen_R;
 
 	LADSPA_Data out_L = 0;
 	LADSPA_Data out_R = 0;
@@ -591,7 +591,7 @@ _init() {
 
 void
 delete_descriptor(LADSPA_Descriptor * descriptor) {
-	unsigned long index;
+	int index;
 	if (descriptor) {
 		free((char *)descriptor->Label);
 		free((char *)descriptor->Name);
